Separate finite element field, value_storage sources
