+++ core/source/finite_element/finite_element_private.h
@@ -68,78 +68,6 @@ Private functions
 -----------------
 */
 
-PROTOTYPE_INDEXED_LIST_STL_IDENTIFIER_CHANGE_FUNCTIONS(FE_field,name);
-
-struct FE_field_info *CREATE(FE_field_info)(struct FE_region *fe_region);
-/*******************************************************************************
-LAST MODIFIED : 2 April 2003
-
-DESCRIPTION :
-Creates a struct FE_field_info with a pointer to <fe_region>.
-Note:
-This should only be called by FE_region functions, and the FE_region must be
-its own master. The returned object is owned by the FE_region.
-It maintains a non-ACCESSed pointer to its owning FE_region which the FE_region
-will clear before it is destroyed. If it becomes necessary to have other owners
-of these objects, the common parts of it and FE_region should be extracted to a
-common object.
-==============================================================================*/
-
-int DESTROY(FE_field_info)(
-	struct FE_field_info **fe_field_info_address);
-/*******************************************************************************
-LAST MODIFIED : 2 April 2003
-
-DESCRIPTION :
-Destroys the FE_field_info at *<field_info_address>. Frees the
-memory for the information and sets <*field_info_address> to NULL.
-==============================================================================*/
-
-PROTOTYPE_OBJECT_FUNCTIONS(FE_field_info);
-
-int FE_field_info_clear_FE_region(struct FE_field_info *field_info);
-/*******************************************************************************
-LAST MODIFIED : 2 April 2003
-
-DESCRIPTION :
-Clears the pointer to FE_region in <field_info>.
-Private function only to be called by destroy_FE_region.
-==============================================================================*/
-
-int FE_field_set_FE_field_info(struct FE_field *fe_field,
-	struct FE_field_info *fe_field_info);
-/*******************************************************************************
-LAST MODIFIED : 2 April 2003
-
-DESCRIPTION :
-Changes the FE_field_info at <fe_field> to <fe_field_info>.
-Private function only to be called by FE_region when merging FE_regions!
-==============================================================================*/
-
-int FE_field_set_indexer_field(struct FE_field *fe_field,
-	struct FE_field *indexer_fe_field);
-/*******************************************************************************
-LAST MODIFIED : 1 May 2003
-
-DESCRIPTION :
-If <fe_field> is already indexed, substitutes <indexer_fe_field>.
-Does not change any of the values currently stored in <fe_field>
-Used to merge indexed fields into different FE_regions; should not be used for
-any other purpose.
-==============================================================================*/
-
-int FE_field_log_FE_field_change(struct FE_field *fe_field,
-	void *fe_field_change_log_void);
-/*******************************************************************************
-LAST MODIFIED : 30 May 2003
-
-DESCRIPTION :
-Logs the field in <fe_field> as RELATED_OBJECT_CHANGED in the
-struct CHANGE_LOG(FE_field) pointed to by <fe_field_change_log_void>.
-???RC Later may wish to allow more than just RELATED_OBJECT_CHANGED, or have
-separate functions for each type.
-==============================================================================*/
-
 /**
  * Describes storage of global values and derivatives for a field at a node.
  * Public only to allow efficient conversion of legacy node DOF indexes.
